requirements.txt
pandas
jinja2
python-dotenv
openpyxl
msoffcrypto-tool
streamlit>=1.32
